VideoForm integration with logic
