+++ RecodoDesktop/RecodoDesktop/Recodo.Desktop.Main.csproj
@@ -7,7 +7,7 @@
 	</PropertyGroup>
 
 	<ItemGroup>
-		<None Remove="camera_preview.png" />
+		<None Remove="camera_preview.jpg" />
 		<None Remove="Icons\bell.png" />
 		<None Remove="Icons\camera.png" />
 		<None Remove="Icons\home.png" />
@@ -37,10 +37,9 @@
 	</ItemGroup>
 
 	<ItemGroup>
-		<Resource Include="camera_preview.png" />
+		<Resource Include="Icons\camera_preview.jpg" />
 		<Resource Include="Icons\bell.png" />
 		<Resource Include="Icons\camera.png" />
-		<Resource Include="Icons\camera_preview.png" />
 		<Resource Include="Icons\home.png" />
 		<Resource Include="Icons\image.png" />
 		<Resource Include="Icons\menu.png" />
